services/http_downloader/requirements.txt
gallery-dl==1.21.1
pika==1.2.0
python-dotenv==0.20.0